sql: fix UB in sqlStrAccumEnlarge()
The patch cast variable `N` to int64 to fix signed integer
overflow. The issue was found by regression test
`sql-luatest/ghs_122_allocations_in_printf_test.lua`, so
an additional test is not required.

Fixes #11176
Part of #10840

NO_CHANGELOG=fix UB
NO_DOC=fix UB
NO_TEST=not required

+++ src/box/sql/printf.c
@@ -829,7 +829,7 @@ sqlStrAccumEnlarge(StrAccum * p, int N)
 		i64 szNew = p->nChar;
 		assert((p->zText == 0
 			|| p->zText == p->zBase) == !isMalloced(p));
-		szNew += N + 1;
+		szNew += (i64)N + 1;
 		if (szNew + p->nChar <= p->mxAlloc) {
 			/* Force exponential buffer size growth as long as it does not overflow,
 			 ** to avoid having to call this routine too often */
